refactor: converter을 이용해 srp 적용
