+++ denstiny/main-server/src/main/java/com/example/domain/dentist/service/CategoryDentistService.java
@@ -5,6 +5,7 @@ import com.example.document.StaticInfoDoc;
 import com.example.domain.dentist.controller.model.CategoryDto;
 import com.example.domain.dentist.controller.model.CategoryLocDto;
 import com.example.domain.dentist.controller.model.DentistDto;
+import com.example.domain.dentist.converter.DentistConverter;
 import com.example.jwt.JWTUtil;
 import com.example.repository.DynamicInfoRepository;
 import com.example.repository.StaticInfoRepository;
@@ -29,8 +30,8 @@ import org.springframework.stereotype.Service;
 public class CategoryDentistService {
 
     private final DynamicInfoRepository dynamicInfoRepository;
-    private final StaticInfoRepository staticInfoRepository;
     private final JWTUtil jwtUtil;
+    private final DentistConverter dentistConverter;
     private final UserRepository userRepository;
 
     public List<DentistDto> categoryDentist(
@@ -47,7 +48,10 @@ public class CategoryDentistService {
             throw new ApiException(ErrorCode.NULL_POINT, "DTO 필드 값이 null입니다.");
         }
 
-        List<DentistDto> dentistDtos = getDentistDtos(category);
+        // 카테고리로 dynamicInfo 정보 조회
+        List<DynamicInfoDoc> dynamicInfoList = dynamicInfoRepository.findByTreatCate(category);
+
+        List<DentistDto> dentistDtos = dentistConverter.toDentistDtos(dynamicInfoList);
 
         // 현재 위치와 병원 간의 거리를 기준으로 병원 정렬
         return dentistDtos.stream()
@@ -78,7 +82,11 @@ public class CategoryDentistService {
         Double longitude = user.getLongitude();
         log.info("{}, {}", latitude, longitude);
 
-        List<DentistDto> dentistDtos = getDentistDtos(category);
+        // 카테고리로 dynamicInfo 정보 조회
+        List<DynamicInfoDoc> dynamicInfoList = dynamicInfoRepository.findByTreatCate(category);
+
+        List<DentistDto> dentistDtos = dentistConverter.toDentistDtos(dynamicInfoList);
+
 
         // 병원 간의 거리를 기준으로 병원 정렬
         List<DentistDto> sortedHospitals = dentistDtos.stream()
@@ -88,44 +96,5 @@ public class CategoryDentistService {
 
         return sortedHospitals;
 
-    }
-
-    private List<DentistDto> getDentistDtos(String category) {
-        // 카테고리로 동적 정보 조회
-        List<DynamicInfoDoc> dynamicInfoList = dynamicInfoRepository.findByTreatCate(category);
-
-        // 동적 정보의 ID 목록으로 정적 정보 한 번에 조회
-        List<String> ids = dynamicInfoList.stream()
-                .map(DynamicInfoDoc::getId)
-                .collect(Collectors.toList());
-
-        Map<String, StaticInfoDoc> staticInfoMap = staticInfoRepository.findAllById(ids).stream()
-                .collect(Collectors.toMap(StaticInfoDoc::getId, staticInfoDoc -> staticInfoDoc));
-
-        List<DentistDto> dentistDtos = dynamicInfoList.stream()
-                .map(dynamicInfo -> {
-                    StaticInfoDoc staticInfo = staticInfoMap.get(dynamicInfo.getId());
-
-                    if (staticInfo == null) return null;  // 정적 정보가 없는 경우 null 반환
-
-                    return DentistDto.builder()
-                            .id(staticInfo.getId())
-                            .name(staticInfo.getName())
-                            .addr(staticInfo.getAddr())
-                            .dong(staticInfo.getDong())
-                            .tele(staticInfo.getTele())
-                            .img(staticInfo.getImg())
-                            .latitude(staticInfo.getLat())
-                            .longitude(staticInfo.getLon())
-                            .score(dynamicInfo.getScore())
-                            .reviewCnt(dynamicInfo.getReviewCnt())
-                            .subwayInfo(staticInfo.getSubwayInfo())
-                            .subwayName(staticInfo.getSubwayName())
-                            .dist(staticInfo.getDist())
-                            .build();
-                })
-//                .filter(Objects::nonNull)
-                .collect(Collectors.toList());
-        return dentistDtos;
     }
 }
